app/server/python/requirements.txt
gunicorn 
Flask 
Flask-RESTful
ffmpeg-python 
pymongo 
Pillow
uuid
flask-cors
python-dotenv
numpy
scipy
scikit-image
matplotlib
pandas
Keras
opencv-python
ipython
tensorflow
tensorflow-addons
tf_slim
opencv-python
opencv-contrib-python
tqdm